[fix] Path 클래스 PathFinder로 네이밍 변경
